requirements.txt
torch>=1.8.0
torchvision>=0.9.0
numpy>=1.19.0
pandas>=1.2.0
opencv-python>=4.5.0
tqdm>=4.60.0
tensorboard>=2.5.0
python-dotenv>=0.19.0
